capa de presentación casi terminada, pendiente el service de room

+++ src/main/java/com/dh/emarket/dao/IDao.java
@@ -5,9 +5,9 @@ import java.util.List;
 public interface IDao<T> {
 
     T save(T t);
-    T findById(Integer id);
+    T findById(Long id);
     void update(T t);
-    void delete(Integer id);
+    void delete(Long id);
     List<T> findAll();
     T findByString(String value);
 }
